esp32-blockchain-dashboard/backend/requirement.txt
flask==2.3.3
flask-cors==4.0.0
pusher==4.1.0
python-dotenv==1.0.0
cryptography==41.0.7
requests==2.31.0